svn path=/branches/cache_manager_rewrite/; revision=15789

+++ reactos/drivers/net/npf/win_bpf.h
@@ -47,7 +47,7 @@
 
 typedef UCHAR u_char;
 typedef USHORT u_short;
-typedef ULONG u_int;
+typedef UINT u_int;
 typedef LONG bpf_int32;
 typedef ULONG bpf_u_int32;
 typedef ULONG u_int32;
